requirements.txt
# transformers @ git+https://github.com/huggingface/transformers.git@a1afecaeb0fb0fc4d80142044b5d1815155ffdcf
transformers
sam3 @ git+https://github.com/Guereak/sam3-apple-silicon.git
accelerate==1.11.0
tokenizers>=0.19.0
huggingface_hub==1.1.5
numpy==1.26.0
torch
torchvision==0.24.1
requests
pillow
dotenv
matplotlib
einops
opencv-python<4.12
pycocotools
tifffile
